+++ plugin/buildSrc/src/main/kotlin/PluginDependencies.kt
@@ -1,7 +1,7 @@
 object PluginVersions {
     val kotlin = "1.4.30"
     val ktlintPlugin = "9.3.0"
-    val ktlint = "0.40.0"
+    val ktlint = "0.41.0"
     val gradlePublishPlugin = "0.12.0"
     val androidPlugin = "4.1.0"
     val semver = "1.1.1"